chore: rename articles.domain
